[chore][pkg/stanza] Cleanup flatten operator files
