+++ pkg/stanza/operator/transformer/flatten/transformer_test.go
